added comments and refactored cube schema file
