+++ backend/wmg/pipeline/summary_cubes/marker_genes.py
@@ -4,7 +4,7 @@ import pandas as pd
 import tiledb
 import numpy as np
 import gc
-from backend.wmg.data.schemas.marker_genes_cube_schema import marker_genes_schema
+from backend.wmg.data.schemas.marker_gene_cube_schema import marker_genes_schema
 from backend.wmg.data.snapshot import CELL_COUNTS_CUBE_NAME, MARKER_GENES_CUBE_NAME
 from backend.wmg.data.utils import create_empty_cube, log_func_runtime
 from backend.wmg.api.calculate_markers import get_markers
